+++ app/controllers/application_controller.rb
@@ -1,14 +1,11 @@
 class ApplicationController < ActionController::Base
+  before_filter :authenticate_user!
+
   include Userstamp
 
   helper :all
-
-  # See ActionController::RequestForgeryProtection for details
-  # Uncomment the :secret if you're not using the cookie session store
-  protect_from_forgery # :secret => '7becd11968ff1efe5b37bfc0a780d63d'
-  
-  include AuthenticatedSystem
-  before_filter :login_required
+  protect_from_forgery # See ActionController::RequestForgeryProtection for details
+  filter_parameter_logging :password, :password_confirmation
 
   layout 'proiel' 
 
@@ -28,5 +25,32 @@ class ApplicationController < ActionController::Base
   # Returns the current page for will_paginate based actions.
   def current_page
     @page ||= params[:page].blank? ? 1 : params[:page].to_i
+  end
+
+  def check_role(role)
+    unless user_signed_in? && @current_user.has_role?(role)
+      if user_signed_in?
+        flash[:error] = 'You do not have permission to access this feature.'
+        redirect_back_or_default('/')
+      else
+        access_denied
+      end
+    end
+  end
+
+  def is_administrator?
+    check_role(:administrator)
+  end
+
+  def is_reviewer?
+    check_role(:reviewer)
+  end
+
+  def is_annotator?
+    check_role(:annotator)
+  end
+
+  def is_reader?
+    check_role(:reader)
   end
 end