+++ config/environments/development.rb
@@ -1,6 +1,3 @@
-# Settings specified here will take precedence over those in config/environment.rb
-config.active_record.observers = :user_observer
-
 # In the development environment your application's code is reloaded on
 # every request.  This slows down response time but is perfect for development
 # since you don't have to restart the webserver when you make code changes.
@@ -18,5 +15,4 @@ config.action_view.debug_rjs                         = true
 config.action_mailer.raise_delivery_errors = true 
 config.action_mailer.delivery_method = :sendmail
 
-SITE_CANONICAL_URL = 'http://foni.uio.no:3001'
-SITE_ADMINISTRATOR_EMAIL = 'mariuslj@ifi.uio.no'
+config.action_mailer.default_url_options = { :host => 'foni.uio.no:3001' }